Add button for Done that works a little bit

+++ parts/delete_todos.php
@@ -1,9 +1,20 @@
 <?php 
+header("Location: ../index.php");
 
-foreach ($_POST as $key => $value) {
-    $statement = $pdo->prepare(
-        "DELETE FROM todo WHERE id = " . $value
-    );
-    $statement->execute();
-    //echo "DELETE FROM todo WHERE id = " . $value;
+require 'database.php';
+
+if(isset($_POST['delete'])){
+    foreach ($_POST as $key => $value) {
+        $statement = $pdo->prepare(
+            "DELETE FROM todo WHERE id = " . $value
+        );
+        $statement->execute();
+    }
+}else if(isset($_POST['done'])){
+    foreach ($_POST as $key => $value) {
+        $statement = $pdo->prepare(
+            "UPDATE todo SET completed = 1 WHERE id = " . $value
+        );
+        $statement->execute();
+    }
 }
